ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
numpy==1.26.4
Pillow==11.0.0
deepface==0.0.93
tf-keras==2.18.0
tensorflow==2.18.0
requests==2.32.3
